requirements.txt
faster-whisper>=1.0
ctranslate2>=4.3.0
huggingface-hub>=0.25.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
pyperclip>=1.8.2
keyboard>=0.13.5
pillow>=10.0.0
pystray>=0.19.4
nvidia-ml-py>=12.0.0
psutil>=5.9.0
plyer>=2.1.0
pywebview>=4.0
av>=12.0.0
onnxruntime>=1.16.0
